Moved session to the separate package
