+++ packages/bitcoin/redux/constants.js
@@ -4,7 +4,6 @@
  */
 
 export const DUCK_BITCOIN = 'bitcoin'
-export const DUCK_SESSION = 'session'
 
 
 export const BITCOIN_CREATE_WALLET = 'BITCOIN/CREATE_WALLET'
