+++ packages/session/package.json
@@ -1,0 +1,4 @@
+{
+  "name": "@chronobank/session",
+  "version": "1.0.0"
+}
