+++ packages/session/redux/initialState.js
