+++ packages/session/redux/reducers.js
